* sysdeps/unix/sysv/linux/aarch64/dl-auxv.h 	(__libc_arm64_midr): Change to int. 	* sysdeps/unix/sysv/linux/aarch64/dl-sysdep.c 	(__libc_arm64_midr): Likewise.

+++ sysdeps/unix/sysv/linux/aarch64/dl-sysdep.c
@@ -1,5 +1,5 @@
 #include "dl-auxv.h"
 
-long long __libc_arm64_midr = -1;
+int __libc_arm64_midr = -1;
 
 #include <sysdeps/unix/sysv/linux/dl-sysdep.c>
